We can postprocess some internal variables and other
especially the average hydrostatic pressure on the REV is calculated and
stored in an output textfile. NB in order to compile this version you
need to compile MGIS on the branch 149 because we need the context class
and methods in MFEM/MGIS. And this is not available yet at this time on
the master of mgis.