Reshaped the search for the maximum stress
We needed to improve the performance for this part, so this commit
address this issue in two manners: first, we adopt a more effective solution of going through the datas of the bubbles,
by sorting first the fetched locations in ascending order of the x
coordinate, and then discarding already the points in which at
least one of the coordinates is alone further than the minimal distance
from the bubble center. On top of it, we use the forall function
provided by the mfem framework, which can allow the speed up of the for
cycle leveraging an omp parallelisation.

Finally, i added the possibility of setting the loading pressure and
minimal distance for bubble fracturing by command line for a better user
experience.